Move to use Shade modules for keystone interaction
We were using the OSA "keystone" module for keystone endpoint, user,
service and role assignments - this requires us to use the OSA plugins
as well, which we shouldn't need. Instead we should just create a
clouds.yaml which is generic to OpenStack in general, and utilise the
existing shade modules to interact with keystone.